+++ java/client/src/main/java/org/signal/libsignal/net/AuthMessagesService.kt
@@ -16,18 +16,21 @@ public class AuthMessagesService(
    * Get an attachment upload form
    *
    * All exceptions are mapped into [RequestResult]; unexpected ones will be treated as
-   * [RequestResult.ApplicationError].
+   * [RequestResult.ApplicationError]. A [UploadTooLargeException] means that the uploadSize was
+   * too large.
    */
-  public fun getUploadForm(): CompletableFuture<RequestResult<UploadForm, Nothing>> =
+  public fun getUploadForm(uploadSize: Long): CompletableFuture<RequestResult<UploadForm, UploadTooLargeException>> =
     try {
+      require(uploadSize >= 0, { "uploadSize ($uploadSize) wasn't >= 0" })
       connection.runWithContextAndConnectionHandles { asyncCtx, conn ->
         Native
           .AuthenticatedChatConnection_get_upload_form(
             asyncCtx,
             conn,
+            uploadSize,
           ).mapWithCancellation(
             onSuccess = { RequestResult.Success(it as UploadForm) },
-            onError = { err -> err.toRequestResult() },
+            onError = { err -> err.toRequestResult<UploadTooLargeException>() },
           )
       }
     } catch (e: Throwable) {